block.rs: Replace ASM-based dbl() with pure-Rust one
Out of paranoia that Rust might produce non-constant-time code for this
codepath, it was previously implemented in terms of the ASM compiler output for
a reference Rust implementation of the function, verified constant-time by
experts.

However, ASM has worries of its own, and carries the risk that using it
might introduce additional vulnerabilities if it is used unsafely.

This commit replaces the ASM implementation of the `dbl()` function with the
pure Rust reference version. So far, the Rust compiler seems good about doing
the right thing with u128s.

In Rust we trust.