Refactor using 'ctr' and 'stream-cipher' crates
The 'ctr' crate provides a generic abstraction for building
`StreamCipher` types out of `BlockCipher`s: counter mode (CTR).

Using this, and going back to a more trait-driven API, we can eliminate
many of the (needless and unwieldy) generic type parameters introduced
in bb7e0d08 (which used the `block-modes` crate for the first time).

Fixing this was the big blocker for another release, so hopefully after
this 0.4 will be nearly ready to release.

This commit deletes all of the FFI code rather than attempting to update
it to the new API. It should likely be replaced by cbindgen, rather than
handrolled.