Add back (off-by-default) "soft-aes" feature
The `aes` crate v0.3 eliminated feature-gated fallback to soft AES.

This adds back a `soft-aes` feature to opt-in to a software
implementation of AES on platforms where a hardware implementation is
not available.

Right now the only supported hardware implementation is AES-NI.
This approach mirrors the target gating used by the `aes` crate,
and therefore unless the `soft-aes` feature is enabled this crate
will not compile on any other architecture besides x86(-64) with
AES-NI target features enabled.