Add back C binding
This re-adds the C FFI/ABI which was removed in 7728e92 (GitHub PR #6).

It was originally removed to simplify that PR, and in the hopes it could
be replaced automatically with cbindgen, but the latter was complicated
by this crate's use of generic-array.

Switching to cbindgen might make more sense after const generics land
and it becomes possible to generate some functions specialized to
particular associated constants.