Use `aes-siv` crate impls; temp remove `alloc` feature
This commit switches the core of the library's implementation to use the
`aes-siv` crate, which is in and of itself an extraction from Miscreant.

Right now the `aead` crate used by `aes-siv` mandates usage of `alloc`.
Unfortunately this means until that's fixed this crate can't support an
optional `alloc` feature, so this commit drops it for now until that can
be addressed.